Use qreal for object and pattern rotation
Be consistent, also with Qt APIs, and avoid (implicit) conversions.

+++ src/core/objects/object.h
@@ -193,11 +193,11 @@ public:
 	
 	/** Rotates the whole object around the center point.
 	 *  The angle must be given in radians. */
-	void rotateAround(MapCoordF center, double angle);
+	void rotateAround(MapCoordF center, qreal angle);
 	
 	/** Rotates the whole object around the center (0, 0).
 	 *  The angle must be given in radians. */
-	void rotate(double angle);
+	void rotate(qreal angle);
 	
 	/** Apply a transformation to all coordinates.
 	 * 
@@ -602,13 +602,13 @@ public:
 	 * Returns the rotation of the object pattern. Only has an effect in
 	 * combination with a symbol interpreting this value.
 	 */
-	float getPatternRotation() const;
+	qreal getPatternRotation() const;
 	
 	/**
 	 * Sets the rotation of the object pattern. Only has an effect in
 	 * combination with a symbol interpreting this value.
 	 */
-	void setPatternRotation(float rotation);
+	void setPatternRotation(qreal rotation);
 	
 	/**
 	 * Returns the origin of the object pattern. Only has an effect in
@@ -902,7 +902,7 @@ private:
 	 * Rotation angle of the object pattern. Only used if the object
 	 * has a symbol which interprets this value.
 	 */
-	float pattern_rotation;
+	qreal pattern_rotation;
 	
 	/**
 	 * Origin shift of the object pattern. Only used if the object
@@ -990,7 +990,7 @@ public:
 	 * error to call setRotation on such an object with an argument other than
 	 * binary 0.
 	 */
-	void setRotation(float new_rotation);
+	void setRotation(qreal new_rotation);
 	
 	/**
 	 * Sets the point object's rotation according to the given vector.
@@ -1001,7 +1001,7 @@ public:
 	 * Returns the point object's rotation (in radians). This is only used
 	 * if the object has a symbol which interprets this value.
 	 */
-	float getRotation() const;
+	qreal getRotation() const;
 	
 	
 	bool intersectsBox(const QRectF& box) const override;
@@ -1009,7 +1009,7 @@ public:
 	
 private:
 	/** The object's rotation (in radians). */
-	float rotation;
+	qreal rotation;
 };
 
 
@@ -1197,7 +1197,7 @@ PathPartVector& PathObject::parts()
 }
 
 inline
-float PathObject::getPatternRotation() const
+qreal PathObject::getPatternRotation() const
 {
 	return pattern_rotation;
 }
@@ -1213,7 +1213,7 @@ MapCoord PathObject::getPatternOrigin() const
 //### PointObject inline code ###
 
 inline
-float PointObject::getRotation() const
+qreal PointObject::getRotation() const
 {
 	return rotation;
 }
