+++ src/core/renderables/renderable_implementation.cpp
@@ -725,7 +725,7 @@ TextRenderable::TextRenderable(const TextSymbol* symbol, const TextObject* text_
 	QTransform t { 1.0, 0.0, 0.0, 1.0, anchor_x, anchor_y };
 	t.scale(scale_factor, scale_factor);
 	
-	auto rotation_rad = qreal(text_object->getRotation());
+	auto rotation_rad = text_object->getRotation();
 	if (!qIsNull(rotation_rad))
 	{
 		rotation = -qRadiansToDegrees(rotation_rad);
