+++ src/core/symbols/point_symbol.h
@@ -76,7 +76,7 @@ public:
 	        ObjectRenderables &output,
 	        RenderableOptions options ) const override;
 	
-	void createRenderablesScaled(MapCoordF coord, float rotation, ObjectRenderables& output, float coord_scale = 1.0f) const;
+	void createRenderablesScaled(MapCoordF coord, qreal rotation, ObjectRenderables& output, float coord_scale = 1.0f) const;
 	
 	void createRenderablesIfCenterInside(MapCoordF point_coord, qreal rotation, const QPainterPath* outline, ObjectRenderables& output) const;
 	void createPrimitivesIfCompletelyInside(MapCoordF point_coord, const QPainterPath* outline, ObjectRenderables& output) const;
